Extract wallet settings into own module
Move the Apple Wallet and Google Wallet configuration handlers and
their validation helpers out of settings.ts into a dedicated
settings-wallets.ts, dropping the apple-wallet and google-wallet
imports that became single-use.
